server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
PyJWT
python-dotenv
Flask-CORS
